// frontend/src/api/generated/models/methodEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `TRANSFER` - TRANSFER
 * * `CASH` - CASH
 * * `CARD` - CARD
 * * `CHECK` - CHECK
 * * `OTHER` - OTHER
 */
export type MethodEnum = (typeof MethodEnum)[keyof typeof MethodEnum];

export const MethodEnum = {
  TRANSFER: "TRANSFER",
  CASH: "CASH",
  CARD: "CARD",
  CHECK: "CHECK",
  OTHER: "OTHER",
} as const;